+++ xxl-job-admin/src/main/java/com/xxl/job/admin/core/model/XxlJobInfo.java
@@ -9,9 +9,9 @@ import java.util.Date;
  */
 public class XxlJobInfo {
 	
-	private int id;				// 主键ID
+	private Long id;				// 主键ID
 	
-	private int jobGroup;		// 执行器主键ID
+	private Long jobGroup;		// 执行器主键ID
 	private String jobDesc;
 	
 	private Date addTime;
@@ -43,19 +43,19 @@ public class XxlJobInfo {
 	private long triggerNextTime;	// 下次调度时间
 
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
-	public int getJobGroup() {
+	public Long getJobGroup() {
 		return jobGroup;
 	}
 
-	public void setJobGroup(int jobGroup) {
+	public void setJobGroup(Long jobGroup) {
 		this.jobGroup = jobGroup;
 	}
